+++ sdk/azidentity/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Breaking Changes
 
-- Instances of `ChainedTokenCredential` will now skip looping through the list of source credentials and re-use the last successful credential on subsequent calls to `GetToken`.
+- Instances of `ChainedTokenCredential` will now skip looping through the list of source credentials and re-use the first successful credential on subsequent calls to `GetToken`.
 
 ### Bugs Fixed