+++ sdk/azidentity/CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ### Breaking Changes
 
-- Instances of `ChainedTokenCredential` will now skip looping through the list of source credentials and re-use the first successful credential on subsequent calls to `GetToken`.
+* Instances of `ChainedTokenCredential` will now skip looping through the list of source credentials and re-use the first successful credential on subsequent calls to `GetToken`.
+  * If the `RetryAllSources` option is set to true, it will not assume the first successful credential should be always used.
 
 ### Bugs Fixed